Rename file's name
