+++ front-end/src/pages/sidebar/NewProducts.jsx
